Use Interface pattern for ColorService
